docs(button): add playground example
