+++ src/components/button/__next__/button.mdx
@@ -1,4 +1,4 @@
-import { Meta, ArgTypes, Canvas } from "@storybook/addon-docs/blocks";
+import { Meta, ArgTypes, Canvas, Controls } from "@storybook/addon-docs/blocks";
 
 import * as ButtonStories from "./button.stories";
 
@@ -33,24 +33,31 @@ Use it to submit a form (Save), to advance to the next step in a process (Next),
 import Button, { type ButtonProps } from "carbon-react/lib/components/button/__next__";
 ```
 
+## Playground
+
+Use this interactive example to explore Button props with Storybook controls.
+
+<Canvas of={ButtonStories.Playground} />
+
+<Controls
+  of={ButtonStories.Playground}
+  include={[
+    "children",
+    "variant",
+    "variantType",
+    "size",
+    "disabled",
+    "fullWidth",
+    "inverse",
+    "noWrap",
+    "type",
+    "href",
+    "target",
+    "rel",
+  ]}
+/>
+
 ## Examples
-
-### Default Button
-
-By default, a `Button` with no additional properties pass will render a default
-primary button, with the `button` HTML type.
-
-<Canvas of={ButtonStories.Default} />
-
-### Button Content
-
-Button content is composable; in other words, it's up to you to decide what you
-want the button to show, and how. In the following example, you can see how to
-implement [Icons](../?path=/story/icon--docs) into buttons. If you have multiple
-child elements, wrap them in a `Fragment` to ensure that the correct styling is
-maintained.
-
-<Canvas of={ButtonStories.ButtonContent} />
 
 ### Click Handler
 
@@ -58,99 +65,6 @@ To make use of non-submit and non-reset buttons, provide an `onClick` handler to
 enable actions to take place when users click the button.
 
 <Canvas of={ButtonStories.ClickHandler} />
-
-### Variants
-
-Buttons come in three variants: `default`, `destructive` and `gradient`. To set your
-desired variant, pass the `variant` prop; if no value is passed, `default` is
-used.
-
-Similarly, variants can be one of up to four variant types: `primary`,
-`secondary`, `tertiary` and `subtle`. Supported variant-type combinations are
-as follows:
-
-| Variant       | Primary | Secondary | Tertiary | Subtle |
-| ------------- | ------- | --------- | -------- | ------ |
-| `default`     | ✅      | ✅        | ✅       | ✅      |
-| `destructive` | ✅      | ✅        | ❌       | ❌      |
-| `gradient`    | ❌      | ✅        | ❌       | ❌      |
-
-By default, the `primary` variant type will be used if no `variantType` property
-is passed
-
-<Canvas of={ButtonStories.Variations} />
-
-### Sizes
-
-Buttons can be one of four sizes: `xs`, `small`, `medium` and `large`. Pass the
-required size with the `size` property.
-
-#### A note on the `xs` size
-
-The `xs` button size is designed for use only in tables where spacing is limited;
-**it is not meant for use beyond that**. XS buttons cannot use the `primary`
-variant type, and will instead default to `secondary` if no variant is specified.
-If `primary` is set, it will be ignored automatically. XS buttons should also use
-only text content (i.e. no icons).
-
-<Canvas of={ButtonStories.Sizes} />
-
-### Disabled
-
-You can disable buttons using the `disabled` property. Before doing that, think
-about whether a disabled button really makes sense for your situation — it's
-best to save this option for times when other approaches won’t quite do the job.
-
-<Canvas of={ButtonStories.Disabled} />
-
-### Full-Width
-
-Passing the `fullWidth` property will allow buttons to use the full horizontal
-space around them.
-
-<Canvas of={ButtonStories.FullWidth} />
-
-### Inverse
-
-Buttons can be marked as `inverse` for scenarios where they require a dark mode
-appearance. Only the `default` variant supports `inverse`; other variants will
-be ignored.
-
-<Canvas of={ButtonStories.Inverse} />
-
-### Loading
-
-Loading state can be applied to buttons by passing the [Loader component](?path=/docs/loader--docs) 
-via the composable `children` prop.
-
-<Canvas of={ButtonStories.Loading} />
-
-### Wrapping Text
-
-If you require button text that does not wrap (due to e.g. limited spacing),
-use the `noWrap` prop to prevent the text from wrapping across multiple
-lines.
-
-<Canvas of={ButtonStories.WrappingText} />
-
-### HTML Button Type
-
-Standard HTML buttons can be one of three types: `button`, `submit` or
-`reset`. This functionality is available via the `type` prop, which
-allows `Button` instances to be used as standard buttons in elements
-such as forms.
-
-<Canvas of={ButtonStories.HTMLButtonType} />
-
-### Button as Link
-
-It is possible to use the `Button` component as a link by passing an `href` prop. 
-When doing so, the component will render as an anchor tag (`<a>`) rather than a button (`<button>`). 
-When using the `Button` component as a link, the `type` prop will be ignored, as it is not 
-relevant for anchor tags. Instead, the `href`, `target` and `rel` props will be used to 
-determine the behaviour of the link.
-
-<Canvas of={ButtonStories.ButtonAsALink} />
 
 ### Programmatic focus
 
@@ -160,8 +74,9 @@ To programmatically set focus on a `Button`, use a ref to the underlying DOM ele
 
 ### Legacy support
 
-To help with migration, this component can also accept the following 
+To help with migration, this component can also accept the following
 [legacy Button](?path=/docs/deprecated-button--docs) props:
+
 - the `buttonType` prop, which maps to the `variant` and `variantType` props as follows:
   - `primary` -> `variant="default"` and `variantType="primary"`
   - `secondary` -> `variant="default"` and `variantType="secondary"`
@@ -178,8 +93,9 @@ To help with migration, this component can also accept the following
 <ArgTypes of={ButtonStories} />
 
 ## Ref
+
 `Button`'s forwarded ref resolves to the underlying DOM element:
 
-| Ref type | Description                                                |
-| -------- | ---------------------------------------------------------- |
+| Ref type | Description                                                      |
+| -------- | ---------------------------------------------------------------- |
 | DOM ref  | `HTMLButtonElement` (or `HTMLAnchorElement` when `href` is set). |
